+++ PyPDF2/_reader.py
@@ -89,7 +89,7 @@ from .generic import (
     TreeObject,
     read_object,
 )
-from .types import OutlinesType, PagemodeType
+from .types import OutlineType, PagemodeType
 from .xmp import XmpInformation
 
 
@@ -644,9 +644,8 @@ class PdfReader:
             # recurse down the tree
             for kid in cast(ArrayObject, tree[PA.KIDS]):
                 self._get_named_destinations(kid.get_object(), retval)
-
         # TABLE 3.33 Entries in a name tree node dictionary (PDF 1.7 specs)
-        if CA.NAMES in tree:
+        elif CA.NAMES in tree:  # KIDS and NAMES are exclusives (PDF 1.7 specs p 162)
             names = cast(DictionaryObject, tree[CA.NAMES])
             for i in range(0, len(names), 2):
                 key = cast(str, names[i].get_object())
@@ -656,7 +655,12 @@ class PdfReader:
                 dest = self._build_destination(key, value)  # type: ignore
                 if dest is not None:
                     retval[key] = dest
-
+        else:  # case where Dests is in root catalog (PDF 1.7 specs, §2 about PDF1.1
+            for k__, v__ in tree.items():
+                val = v__.get_object()
+                dest = self._build_destination(k__, val)
+                if dest is not None:
+                    retval[k__] = dest
         return retval
 
     def getNamedDestinations(
@@ -673,19 +677,30 @@ class PdfReader:
         return self._get_named_destinations(tree, retval)
 
     @property
-    def outlines(self) -> OutlinesType:
+    def outline(self) -> OutlineType:
         """
-        Read-only property for outlines present in the document.
+        Read-only property for the outline (i.e., a collection of 'outline items'
+        which are also known as 'bookmarks') present in the document.
 
         :return: a nested list of :class:`Destinations<PyPDF2.generic.Destination>`.
         """
-        return self._get_outlines()
+        return self._get_outline()
 
-    def _get_outlines(
-        self, node: Optional[DictionaryObject] = None, outlines: Optional[Any] = None
-    ) -> OutlinesType:
-        if outlines is None:
-            outlines = []
+    @property
+    def outlines(self) -> OutlineType:
+        """
+        .. deprecated:: 2.9.0
+
+            Use :py:attr:`outline` instead.
+        """
+        deprecate_with_replacement("outlines", "outline")
+        return self.outline
+
+    def _get_outline(
+        self, node: Optional[DictionaryObject] = None, outline: Optional[Any] = None
+    ) -> OutlineType:
+        if outline is None:
+            outline = []
             catalog = cast(DictionaryObject, self.trailer[TK.ROOT])
 
             # get the outline dictionary and named destinations
@@ -695,11 +710,11 @@ class PdfReader:
                 except PdfReadError:
                     # this occurs if the /Outlines object reference is incorrect
                     # for an example of such a file, see https://unglueit-files.s3.amazonaws.com/ebf/7552c42e9280b4476e59e77acc0bc812.pdf
-                    # so continue to load the file without the Bookmarks
-                    return outlines
+                    # so continue to load the file without the Outlines
+                    return outline
 
                 if isinstance(lines, NullObject):
-                    return outlines
+                    return outline
 
                 # TABLE 8.3 Entries in the outline dictionary
                 if lines is not None and "/First" in lines:
@@ -707,37 +722,37 @@ class PdfReader:
             self._namedDests = self._get_named_destinations()
 
         if node is None:
-            return outlines
+            return outline
 
-        # see if there are any more outlines
+        # see if there are any more outline items
         while True:
-            outline = self._build_outline(node)
-            if outline:
-                outlines.append(outline)
+            outline_obj = self._build_outline_item(node)
+            if outline_obj:
+                outline.append(outline_obj)
 
-            # check for sub-outlines
+            # check for sub-outline
             if "/First" in node:
-                sub_outlines: List[Any] = []
-                self._get_outlines(cast(DictionaryObject, node["/First"]), sub_outlines)
-                if sub_outlines:
-                    outlines.append(sub_outlines)
+                sub_outline: List[Any] = []
+                self._get_outline(cast(DictionaryObject, node["/First"]), sub_outline)
+                if sub_outline:
+                    outline.append(sub_outline)
 
             if "/Next" not in node:
                 break
             node = cast(DictionaryObject, node["/Next"])
 
-        return outlines
+        return outline
 
     def getOutlines(
-        self, node: Optional[DictionaryObject] = None, outlines: Optional[Any] = None
-    ) -> OutlinesType:  # pragma: no cover
+        self, node: Optional[DictionaryObject] = None, outline: Optional[Any] = None
+    ) -> OutlineType:  # pragma: no cover
         """
         .. deprecated:: 1.28.0
 
-            Use :py:attr:`outlines` instead.
+            Use :py:attr:`outline` instead.
         """
-        deprecate_with_replacement("getOutlines", "outlines")
-        return self._get_outlines(node, outlines)
+        deprecate_with_replacement("getOutlines", "outline")
+        return self._get_outline(node, outline)
 
     def _get_page_number_by_indirect(
         self, indirect_ref: Union[None, int, NullObject, IndirectObject]
@@ -805,7 +820,7 @@ class PdfReader:
         array: List[Union[NumberObject, IndirectObject, NullObject, DictionaryObject]],
     ) -> Destination:
         page, typ = None, None
-        # handle outlines with missing or invalid destination
+        # handle outline items with missing or invalid destination
         if (
             isinstance(array, (type(None), NullObject))
             or (isinstance(array, ArrayObject) and len(array) == 0)
@@ -831,8 +846,8 @@ class PdfReader:
                     title, indirect_ref, TextStringObject("/Fit")  # type: ignore
                 )
 
-    def _build_outline(self, node: DictionaryObject) -> Optional[Destination]:
-        dest, title, outline = None, None, None
+    def _build_outline_item(self, node: DictionaryObject) -> Optional[Destination]:
+        dest, title, outline_item = None, None, None
 
         # title required for valid outline
         # PDF Reference 1.7: TABLE 8.4 Entries in an outline item dictionary
@@ -857,40 +872,40 @@ class PdfReader:
                 dest = dest["/D"]
 
         if isinstance(dest, ArrayObject):
-            outline = self._build_destination(title, dest)  # type: ignore
+            outline_item = self._build_destination(title, dest)  # type: ignore
         elif isinstance(dest, str):
             # named destination, addresses NameObject Issue #193
             try:
-                outline = self._build_destination(
+                outline_item = self._build_destination(
                     title, self._namedDests[dest].dest_array
                 )
             except KeyError:
                 # named destination not found in Name Dict
-                outline = self._build_destination(title, None)
+                outline_item = self._build_destination(title, None)
         elif isinstance(dest, type(None)):
-            # outline not required to have destination or action
+            # outline item not required to have destination or action
             # PDFv1.7 Table 153
-            outline = self._build_destination(title, dest)  # type: ignore
+            outline_item = self._build_destination(title, dest)  # type: ignore
         else:
             if self.strict:
                 raise PdfReadError(f"Unexpected destination {dest!r}")
-            outline = self._build_destination(title, None)  # type: ignore
+            outline_item = self._build_destination(title, None)  # type: ignore
 
-        # if outline created, add color, format, and child count if present
-        if outline:
+        # if outline item created, add color, format, and child count if present
+        if outline_item:
             if "/C" in node:
-                # Color of outline in (R, G, B) with values ranging 0.0-1.0
-                outline[NameObject("/C")] = ArrayObject(FloatObject(c) for c in node["/C"])  # type: ignore
+                # Color of outline item font in (R, G, B) with values ranging 0.0-1.0
+                outline_item[NameObject("/C")] = ArrayObject(FloatObject(c) for c in node["/C"])  # type: ignore
             if "/F" in node:
                 # specifies style characteristics bold and/or italic
                 # 1=italic, 2=bold, 3=both
-                outline[NameObject("/F")] = node["/F"]
+                outline_item[NameObject("/F")] = node["/F"]
             if "/Count" in node:
                 # absolute value = num. visible children
                 # positive = open/unfolded, negative = closed/folded
-                outline[NameObject("/Count")] = node["/Count"]
+                outline_item[NameObject("/Count")] = node["/Count"]
 
-        return outline
+        return outline_item
 
     @property
     def pages(self) -> _VirtualList:
@@ -957,9 +972,9 @@ class PdfReader:
            :widths: 50 200
 
            * - /UseNone
-             - Do not show outlines or thumbnails panels
+             - Do not show outline or thumbnails panels
            * - /UseOutlines
-             - Show outlines (aka bookmarks) panel
+             - Show outline (aka bookmarks) panel
            * - /UseThumbs
              - Show page thumbnails panel
            * - /FullScreen
@@ -1303,7 +1318,7 @@ class PdfReader:
             stream.seek(0, os.SEEK_END)
 
     def _find_eof_marker(self, stream: StreamType) -> None:
-        last_mb = stream.tell() - 1024 * 1024 + 1  # offset of last MB of stream
+        last_mb = 8  # to parse whole file
         line = b""
         while line[:5] != b"%%EOF":
             if stream.tell() < last_mb: