+++ tests/test_reader.py
@@ -11,11 +11,7 @@ from PyPDF2._reader import convert_to_int, convertToInt
 from PyPDF2.constants import ImageAttributes as IA
 from PyPDF2.constants import PageAttributes as PG
 from PyPDF2.constants import Ressources as RES
-from PyPDF2.errors import (
-    STREAM_TRUNCATED_PREMATURELY,
-    PdfReadError,
-    PdfReadWarning,
-)
+from PyPDF2.errors import PdfReadError, PdfReadWarning
 from PyPDF2.filters import _xobj_to_image
 from PyPDF2.generic import Destination
 
@@ -144,10 +140,10 @@ def test_get_attachments(src):
         (os.path.join(RESOURCE_ROOT, "crazyones.pdf"), 0),
     ],
 )
-def test_get_outlines(src, outline_elements):
+def test_get_outline(src, outline_elements):
     reader = PdfReader(src)
-    outlines = reader._get_outlines()
-    assert len(outlines) == outline_elements
+    outline = reader.outline
+    assert len(outline) == outline_elements
 
 
 @pytest.mark.parametrize(
@@ -434,7 +430,9 @@ def test_read_malformed_header():
 def test_read_malformed_body():
     with pytest.raises(PdfReadError) as exc:
         PdfReader(io.BytesIO(b"%PDF-"), strict=True)
-    assert exc.value.args[0] == STREAM_TRUNCATED_PREMATURELY
+    assert (
+        exc.value.args[0] == "EOF marker not found"
+    )  # used to be:STREAM_TRUNCATED_PREMATURELY
 
 
 def test_read_prev_0_trailer():
@@ -573,10 +571,10 @@ def test_read_encrypted_without_decryption():
 def test_get_destination_page_number():
     src = os.path.join(RESOURCE_ROOT, "pdflatex-outline.pdf")
     reader = PdfReader(src)
-    outlines = reader._get_outlines()
-    for outline in outlines:
-        if not isinstance(outline, list):
-            reader.get_destination_page_number(outline)
+    outline = reader.outline
+    for outline_item in outline:
+        if not isinstance(outline_item, list):
+            reader.get_destination_page_number(outline_item)
 
 
 def test_do_not_get_stuck_on_large_files_without_start_xref():
@@ -609,7 +607,7 @@ def test_decrypt_when_no_id():
 
 def test_reader_properties():
     reader = PdfReader(os.path.join(RESOURCE_ROOT, "crazyones.pdf"))
-    assert reader.outlines == []
+    assert reader.outline == []
     assert len(reader.pages) == 1
     assert reader.page_layout is None
     assert reader.page_mode is None
@@ -624,17 +622,17 @@ def test_issue604(caplog, strict):
     """Test with invalid destinations"""  # todo
     with open(os.path.join(RESOURCE_ROOT, "issue-604.pdf"), "rb") as f:
         pdf = None
-        bookmarks = None
+        outline = None
         if strict:
             pdf = PdfReader(f, strict=strict)
             with pytest.raises(PdfReadError) as exc, pytest.warns(PdfReadWarning):
-                bookmarks = pdf._get_outlines()
+                outline = pdf.outline
             if "Unknown Destination" not in exc.value.args[0]:
                 raise Exception("Expected exception not raised")
-            return  # bookmarks not correct
+            return  # outline is not correct
         else:
             pdf = PdfReader(f, strict=strict)
-            bookmarks = pdf._get_outlines()
+            outline = pdf.outline
             msg = (
                 "WARNING  PyPDF2._reader:_utils.py:364 "
                 "Unknown destination: ms_Thyroid_2_2020_071520_watermarked.pdf [0, 1]\n"
@@ -651,8 +649,8 @@ def test_issue604(caplog, strict):
         out = []
 
         # b can be destination or a list:preferred to just print them
-        for b in bookmarks:
-            out.append(get_dest_pages(b))
+        for oi in outline:
+            out.append(get_dest_pages(oi))
 
 
 def test_decode_permissions():
@@ -908,33 +906,33 @@ def test_outline_color():
     url = "https://corpora.tika.apache.org/base/docs/govdocs1/924/924546.pdf"
     name = "tika-924546.pdf"
     reader = PdfReader(BytesIO(get_pdf_from_url(url, name=name)))
-    assert reader.outlines[0].color == [0, 0, 1]
+    assert reader.outline[0].color == [0, 0, 1]
 
 
 def test_outline_font_format():
     url = "https://corpora.tika.apache.org/base/docs/govdocs1/924/924546.pdf"
     name = "tika-924546.pdf"
     reader = PdfReader(BytesIO(get_pdf_from_url(url, name=name)))
-    assert reader.outlines[0].font_format == 2
+    assert reader.outline[0].font_format == 2
 
 
-def get_outlines_property(outlines, attribute_name: str):
+def get_outline_property(outline, attribute_name: str):
     results = []
-    if isinstance(outlines, list):
-        for outline in outlines:
-            if isinstance(outline, Destination):
-                results.append(getattr(outline, attribute_name))
+    if isinstance(outline, list):
+        for outline_item in outline:
+            if isinstance(outline_item, Destination):
+                results.append(getattr(outline_item, attribute_name))
             else:
-                results.append(get_outlines_property(outline, attribute_name))
+                results.append(get_outline_property(outline_item, attribute_name))
     else:
-        raise ValueError(f"got {type(outlines)}")
+        raise ValueError(f"got {type(outline)}")
     return results
 
 
 def test_outline_title_issue_1121():
     reader = PdfReader(EXTERNAL_ROOT / "014-outlines/mistitled_outlines_example.pdf")
 
-    assert get_outlines_property(reader.outlines, "title") == [
+    assert get_outline_property(reader.outline, "title") == [
         "First",
         [
             "Second",
@@ -980,7 +978,7 @@ def test_outline_title_issue_1121():
 def test_outline_count():
     reader = PdfReader(EXTERNAL_ROOT / "014-outlines/mistitled_outlines_example.pdf")
 
-    assert get_outlines_property(reader.outlines, "outline_count") == [
+    assert get_outline_property(reader.outline, "outline_count") == [
         5,
         [
             None,
@@ -1028,8 +1026,23 @@ def test_outline_missing_title():
         os.path.join(RESOURCE_ROOT, "outline-without-title.pdf"), strict=True
     )
     with pytest.raises(PdfReadError) as exc:
-        reader.outlines
+        reader.outline
     assert exc.value.args[0].startswith("Outline Entry Missing /Title attribute:")
+
+
+def test_named_destination():
+    # 1st case : the named_dest are stored directly as a dictionnary, PDF1.1 style
+    url = "https://github.com/py-pdf/PyPDF2/files/9197028/lorem_ipsum.pdf"
+    name = "lorem_ipsum.pdf"
+    reader = PdfReader(BytesIO(get_pdf_from_url(url, name=name)))
+    assert len(reader.named_destinations) > 0
+    # 2nd case : Dest below names and with Kids...
+    url = "https://opensource.adobe.com/dc-acrobat-sdk-docs/standards/pdfstandards/pdf/PDF32000_2008.pdf"
+    name = "PDF32000_2008.pdf"
+    reader = PdfReader(BytesIO(get_pdf_from_url(url, name=name)))
+    assert len(reader.named_destinations) > 0
+    # 3nd case : Dests with Name tree
+    # TODO : case to be added
 
 
 def test_outline_with_missing_named_destination():
@@ -1037,21 +1050,31 @@ def test_outline_with_missing_named_destination():
     name = "tika-913678.pdf"
     reader = PdfReader(BytesIO(get_pdf_from_url(url, name=name)))
     # outline items in document reference a named destination that is not defined
-    assert reader.outlines[1][0].title.startswith("Report for 2002AZ3B: Microbial")
+    assert reader.outline[1][0].title.startswith("Report for 2002AZ3B: Microbial")
 
 
 def test_outline_with_empty_action():
     url = "https://corpora.tika.apache.org/base/docs/govdocs1/924/924546.pdf"
     name = "tika-924546.pdf"
     reader = PdfReader(BytesIO(get_pdf_from_url(url, name=name)))
-    # outline (entitled Tables and Figures) utilize an empty action (/A)
+    # outline items (entitled Tables and Figures) utilize an empty action (/A)
     # that has no type or destination
-    assert reader.outlines[-4].title == "Tables"
+    assert reader.outline[-4].title == "Tables"
 
 
-def test_outlines_with_invalid_destinations():
+def test_outline_with_invalid_destinations():
     reader = PdfReader(
         os.path.join(RESOURCE_ROOT, "outlines-with-invalid-destinations.pdf")
     )
-    # contains 9 outlines, 6 with invalid destinations caused by different malformations
-    assert len(reader.outlines) == 9
+    # contains 9 outline items, 6 with invalid destinations caused by different malformations
+    assert len(reader.outline) == 9
+
+
+def test_PdfReaderMultipleDefinitions():
+    # iss325
+    url = "https://github.com/py-pdf/PyPDF2/files/9176644/multipledefs.pdf"
+    name = "multipledefs.pdf"
+    reader = PdfReader(BytesIO(get_pdf_from_url(url, name=name)))
+    with pytest.warns(PdfReadWarning) as w:
+        reader.pages[0].extract_text()
+    assert len(w) == 1